requirement.txt
asgiref==3.7.2
crispy-bootstrap4==2023.1
crispy-bootstrap5==2023.10
Django==5.0.1
django-crispy-forms==2.1
django-filter==23.5
djangorestframework==3.14.0
psycopg2==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.4
